requirements.txt
fastapi==0.109.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.6.1
python-dotenv==1.0.0
uvicorn==0.27.0.post1